fix: 프로젝션시 생성자 타입이 맞지 않는 오류 수정

+++ src/main/java/com/ipillgood/server/domain/review/repository/ProductReviewProjection.java
@@ -4,6 +4,6 @@ public class ProductReviewProjection {
 
     public record ReviewSummary(
             Double ratingAverage,
-            Integer reviewCount
+            Long reviewCount
     ) {}
 }
